requirements.txt
Django==5.0.3
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
djangorestframework==3.15.0
python-dotenv==1.0.1
pytz==2024.1
# sqlparse==0.4.4
# mysqlclient==2.2.4
whitenoise==5.2.0
packaging~=24.0
# amqp==5.2.0
# billiard==4.2.0
# celery==5.3.6
# django-bootstrap3==23.6
# django-celery-beat==2.6.0
# django-celery-results==2.5.1
requests==2.31.0
# pillow==10.3.0
# ldap3==2.9.1
# openpyxl==3.1.5
pandas==2.2.3